+++ src/test/NoArgConstructor.java
@@ -1,0 +1,9 @@
+/** This class's constructor has zero formal parameters. */
+class NoArgConstructor {
+  NoArgConstructor() {}
+}
+
+/** This class's constructor has a formal parameter. */
+class HasArgConstructor {
+  HasArgConstructor(int i) {}
+}